[Bugs] Fixed floor to height conversion. Created test cases

+++ app/src/main/java/edu/udmercy/accesspointlocater/features/create/repositories/BuildingImageRepositoryImpl.kt
@@ -27,8 +27,8 @@ class BuildingImageRepositoryImpl(private val appContext: Context): BuildingImag
         return buildingImageRepo.getFloorCount(uuid)
     }
 
-    override fun getFloorHeights(uuid: String, floorCounts: Int): List<Float> {
-        val floorHeights = mutableListOf<Float>()
+    override fun getFloorHeights(uuid: String, floorCounts: Int): List<Double> {
+        val floorHeights = mutableListOf<Double>()
         for (i in 0 until floorCounts) {
             floorHeights.add(buildingImageRepo.getFloorHeight(uuid, i))
         }
